Add files via upload
I analyzed a dataset of 80 cereal types from seven manufacturers containing the data about the content in nutrient groups (like fat or proteins) and elements (like potassium). 

I answered the following questions:

What is the average content of each nutrient and element per manufacturer?
What is the number of calories per ounce for each product?
What is the average rating per manufacturer? Is it somehow connected to average nutrient content?
Which manufacturer possesses the best shelf location?
What is the nutritional value of each cereal according to protein, fat, and carbohydrate data.